all test cases pass

+++ include/uint256_t.h
@@ -62,7 +62,7 @@ class uint256_t
     CUDA_CALLABLE_MEMBER uint256_t operator&( const uint256_t& comp ) const;
     CUDA_CALLABLE_MEMBER uint256_t operator|( const uint256_t& comp ) const;
     CUDA_CALLABLE_MEMBER uint256_t operator^( const uint256_t& comp ) const;
-    CUDA_CALLABLE_MEMBER std::uint8_t at( int loc );
+    CUDA_CALLABLE_MEMBER std::uint32_t at( int loc );
     CUDA_CALLABLE_MEMBER uint256_t operator<<( int shift ) const;
     CUDA_CALLABLE_MEMBER uint256_t operator>>( int shift ) const;
     CUDA_CALLABLE_MEMBER uint256_t operator~() const;
